pull common code between FwDataMiniLcmHub.cs and CrdtMiniLcmApiHub.cs into MiniLcmApiHubBase.cs
