+++ backend/FwLite/LocalWebApp/Hubs/CrdtMiniLcmApiHub.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNetCore.Http.Json;
-using Microsoft.AspNetCore.SignalR;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.AspNetCore.SignalR;
 using MiniLcm;
 using SystemTextJsonPatch;
 
@@ -8,9 +6,8 @@ namespace LocalWebApp.Hubs;
 
 public class CrdtMiniLcmApiHub(
     ILexboxApi lexboxApi,
-    IOptions<JsonOptions> jsonOptions,
     BackgroundSyncService backgroundSyncService,
-    SyncService syncService) : Hub<ILexboxHubClient>
+    SyncService syncService) : MiniLcmApiHubBase(lexboxApi)
 {
     public const string ProjectRouteKey = "project";
     public override async Task OnConnectedAsync()
@@ -18,123 +15,70 @@ public class CrdtMiniLcmApiHub(
         await syncService.ExecuteSync();
     }
 
-    public async Task<WritingSystems> GetWritingSystems()
+    public override async Task<WritingSystem> CreateWritingSystem(WritingSystemType type, WritingSystem writingSystem)
     {
-        return await lexboxApi.GetWritingSystems();
-    }
-
-    public async Task<WritingSystem> CreateWritingSystem(WritingSystemType type, WritingSystem writingSystem)
-    {
-        var newWritingSystem = await lexboxApi.CreateWritingSystem(type, writingSystem);
+        var newWritingSystem = await base.CreateWritingSystem(type, writingSystem);
         backgroundSyncService.TriggerSync();
         return newWritingSystem;
     }
 
-    public async Task<WritingSystem> UpdateWritingSystem(WritingSystemId id, WritingSystemType type, JsonPatchDocument<WritingSystem> update)
+    public override async Task<WritingSystem> UpdateWritingSystem(WritingSystemId id, WritingSystemType type, JsonPatchDocument<WritingSystem> update)
     {
-        var writingSystem = await lexboxApi.UpdateWritingSystem(id, type, new JsonPatchUpdateInput<WritingSystem>(update));
+        var writingSystem = await base.UpdateWritingSystem(id, type, update);
         backgroundSyncService.TriggerSync();
         return writingSystem;
     }
 
-    public IAsyncEnumerable<PartOfSpeech> GetPartsOfSpeech()
+    public override async Task<Entry> CreateEntry(Entry entry)
     {
-        return lexboxApi.GetPartsOfSpeech();
-    }
-
-    public IAsyncEnumerable<SemanticDomain> GetSemanticDomains()
-    {
-        return lexboxApi.GetSemanticDomains();
-    }
-
-    public IAsyncEnumerable<Entry> GetEntriesForExemplar(string exemplar, QueryOptions? options = null)
-    {
-        throw new NotImplementedException();
-    }
-
-    public IAsyncEnumerable<Entry> GetEntries(QueryOptions? options = null)
-    {
-        return lexboxApi.GetEntries(options);
-    }
-
-    public IAsyncEnumerable<Entry> SearchEntries(string query, QueryOptions? options = null)
-    {
-        return lexboxApi.SearchEntries(query, options);
-    }
-
-    public async Task<Entry?> GetEntry(Guid id)
-    {
-        return await lexboxApi.GetEntry(id);
-    }
-
-    public async Task<Entry> CreateEntry(Entry entry)
-    {
-        var newEntry = await lexboxApi.CreateEntry(entry);
+        var newEntry = await base.CreateEntry(entry);
         await NotifyEntryUpdated(newEntry);
         return newEntry;
     }
 
-    public async Task<Entry> UpdateEntry(Guid id, JsonPatchDocument<Entry> update)
+    public override async Task<Entry> UpdateEntry(Guid id, JsonPatchDocument<Entry> update)
     {
-        var entry = await lexboxApi.UpdateEntry(id, new JsonPatchUpdateInput<Entry>(update));
+        var entry = await base.UpdateEntry(id, update);
         await NotifyEntryUpdated(entry);
         return entry;
     }
 
-    public async Task DeleteEntry(Guid id)
+    public override async Task<Sense> CreateSense(Guid entryId, Sense sense)
     {
-        await lexboxApi.DeleteEntry(id);
-    }
-
-    public async Task<Sense> CreateSense(Guid entryId, Sense sense)
-    {
-        var createdSense = await lexboxApi.CreateSense(entryId, sense);
+        var createdSense = await base.CreateSense(entryId, sense);
         backgroundSyncService.TriggerSync();
         return createdSense;
     }
 
-    public async Task<Sense> UpdateSense(Guid entryId, Guid senseId, JsonPatchDocument<Sense> update)
+    public override async Task<Sense> UpdateSense(Guid entryId, Guid senseId, JsonPatchDocument<Sense> update)
     {
-        var sense = await lexboxApi.UpdateSense(entryId, senseId, new JsonPatchUpdateInput<Sense>(update));
+        var sense = await base.UpdateSense(entryId, senseId, update);
         backgroundSyncService.TriggerSync();
         return sense;
     }
 
-    public async Task DeleteSense(Guid entryId, Guid senseId)
-    {
-        await lexboxApi.DeleteSense(entryId, senseId);
-    }
-
-    public async Task<ExampleSentence> CreateExampleSentence(Guid entryId,
+    public override async Task<ExampleSentence> CreateExampleSentence(Guid entryId,
         Guid senseId,
         ExampleSentence exampleSentence)
     {
-        var createdSentence = await lexboxApi.CreateExampleSentence(entryId, senseId, exampleSentence);
+        var createdSentence = await base.CreateExampleSentence(entryId, senseId, exampleSentence);
         backgroundSyncService.TriggerSync();
         return createdSentence;
     }
 
-    public async Task<ExampleSentence> UpdateExampleSentence(Guid entryId,
+    public override async Task<ExampleSentence> UpdateExampleSentence(Guid entryId,
         Guid senseId,
         Guid exampleSentenceId,
         JsonPatchDocument<ExampleSentence> update)
     {
-        var sentence = await lexboxApi.UpdateExampleSentence(entryId,
-            senseId,
-            exampleSentenceId,
-            new JsonPatchUpdateInput<ExampleSentence>(update));
+        var sentence = await base.UpdateExampleSentence(entryId, senseId, exampleSentenceId, update);
         backgroundSyncService.TriggerSync();
         return sentence;
     }
 
-    public async Task DeleteExampleSentence(Guid entryId, Guid senseId, Guid exampleSentenceId)
+    protected override async Task NotifyEntryUpdated(Entry entry)
     {
-        await lexboxApi.DeleteExampleSentence(entryId, senseId, exampleSentenceId);
-    }
-
-    private async Task NotifyEntryUpdated(Entry entry)
-    {
-        await Clients.Others.OnEntryUpdated(entry);
+        await base.NotifyEntryUpdated(entry);
         backgroundSyncService.TriggerSync();
     }
 }
